requirements.txt
flask
flask-socketio
python-socketio
eventlet
discord.py
requests
